experiments/voice-triage-web/requirements.txt
pipecat-ai[daily,openai,silero]>=0.0.40
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
jinja2>=3.1.0
